% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_DE.R
\name{batch_de}
\alias{batch_de}
\title{Produce differential expression analysis in batch}
\usage{
batch_de(
  de_infos,
  txi,
  design,
  outdir = NULL,
  r_objects = NULL,
  force = FALSE,
  cores = 1
)
}
\arguments{
\item{de_infos}{A csv file or a \code{data.frame} describing the DE analysis
to perform.}

\item{txi}{The txi object returned by the import_kallisto function.}

\item{design}{A csv file of a \code{data.frame} describing the groups for
the comparisons.}

\item{outdir}{The directory where to save DE analysis in csv format. If
\code{NULL}, the results won't be saved as csv The files will be saved as
<outdir>/<id_de>.csv. Default: \code{NULL}.}

\item{r_objects}{The directory where to save dds in rds format. If
\code{NULL}, the dds won't be saved as rds files. The files will be saved as
<r_objects>/<id_de>.rds. Default: \code{NULL}.}

\item{force}{Should the files be re-created if they already exists? Default:
\code{FALSE}.}

\item{cores}{Number of cores for the DE analysis. Default: 1}
}
\value{
Invisibly returns a \code{list} of all the DE results.
}
\description{
The goal of this function is to take a csv file or a \code{data.frame}
describing all the differential expression (DE) analysis to produce and
launch them in batches.
}
\details{
If the \code{r_objects} param is used, the dds object will be saved and
won't be recalculated if the function is called again, unless the
\code{force} param is set to \code{TRUE}.

The table may contain the following columns, in no specific order:
  * id_de: unique identifier for this DE analysis. Mandatory.
  * group: column from the design file to define the group for the current
  DE analysis. Mandatory.
  * contrast_1: The first group for the DE analysis. Must be present in the
  group column. Mandatory.
  * contrast_2: The first group for the DE analysis. Must be present in the
  group column. Mandatory.
  * formula: The formula for the de_analysis. Default: ~ group
  * filter: Minimum mean number of reads across all samples to be part of
  the DE analysis
  * count_matrix: The count matrix to use for the differential analysis.
  Will use the \code{DESeq2::DESeqDataSetFromMatrix} instead of the
  \code{DESeq2::DESeqDataSetFromTximport} function, so will work even if txi
  object is incomplete (i.e.: length matrix is missing). Default: NA
}
\examples{
\dontrun{
de_infos <- get_demo_de_infos_file()
txi <- get_demo_txi()
design <- get_demo_design()
de_list <- batch_de(de_infos, txi, design)
}

}
