% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txi.R
\name{create_dummy_txi}
\alias{create_dummy_txi}
\title{Create a dummy txi objects}
\usage{
create_dummy_txi(tables, txOut = FALSE)
}
\arguments{
\item{tables}{A \code{list} of the tables to add to the \code{dummy_txi}
object. Must contains at \code{counts} (raw counts) or \code{abundance}
(usually TPM). The \code{anno} is also mandatory.}

\item{txOut}{Are the counts corresponding to transcripts abundance? Default:
\code{FALSE}.}
}
\value{
The \code{dummy_txi} object.
}
\description{
This function will create a dummy txi object that will have placeholders for
some of the tables.
}
\details{
This function is useful if you only have access to raw counts or abundance
tables. Otherwise, it is recommended to use the \code{import_kallisto}
function as it will create a complete and valid \code{txi} object.

The \code{txi} will contains extra elements to specify how it was created
and will provide extra checks to avoid using it incorrectly with the other
functions. For instance, to use the \code{produce_pca_df} function, the
\code{txi} must have a valid \code{abundance} table. If the \code{dummy_txi}
was created without such a table, it will return an error message to avoid
producing an invalid result.

You have to provide at least the raw counts or the abundance table **and**
the corresponding annotation in the same format as the one produced by the
\code{anno:prepare_anno} function.
}
\examples{
txi <- get_demo_txi()

lst <- list()
lst$counts <- txi$counts
lst$abundance <- txi$abundance
lst$anno <- txi$anno

# We then create a dummy txi object
dummy_txi <- create_dummy_txi(txi)
# This dummy_txi contains a placeholder for the length matrix

}
