% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{deseq2_analysis}
\alias{deseq2_analysis}
\title{DESeq2 analysis}
\usage{
deseq2_analysis(txi, design, formula, filter = 2, count_matrix = NULL, ...)
}
\arguments{
\item{txi}{The txi object returned by the import_kallisto function.}

\item{design}{The experimental design (see ?DESeqDataSetFromTximport).}

\item{formula}{The design formula in data.frame format (see
?DESeqDataSetFromTximport).}

\item{filter}{The minimum number of reads detected for a feature across all
samples. Default: 2}

\item{count_matrix}{Use an alternative count matrix to use for the
differential analysis instead of \code{txi$counts}. Will use the
\code{DESeq2::DESeqDataSetFromMatrix} instead of the
\code{DESeq2::DESeqDataSetFromTximport} function, so will work even if txi
object is incomplete (i.e.: length matrix is missing).  Default: \code{NA}.}

\item{...}{Extra param for the DESeq2::DESeq function}
}
\value{
A DESeqDataSet object.
}
\description{
DESeq2 analysis
}
\examples{
txi <- get_demo_txi()
design <- get_demo_design()
dds <- deseq2_analysis(txi, design, ~ group)
de <- DESeq2::results(dds, contrast = c("group", "A", "B"))

}
