% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txi.R
\name{filter_txi}
\alias{filter_txi}
\title{Filter out txi object}
\usage{
filter_txi(txi, samples)
}
\arguments{
\item{txi}{The txi object returned by the import_kallisto function.}

\item{samples}{The list of samples to keep.}
}
\value{
The filter txi object.
}
\description{
This function allow to filter the txi object by taking a list of sample to
keep. All the expected matrices will be filtered (counts, abundance, fpkm,
ruvg_counts and combat_counts.
}
\examples{
txi <- get_demo_txi() # sample names are "a", "b", "c" and "d"
txi_filter <- filter_txi(txi, c("a", "c")) # Only keep "a" and "c"

}
