% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odysse.R
\name{format_de_odysse}
\alias{format_de_odysse}
\title{Prepare formated DE table in odysse format.}
\usage{
format_de_odysse(de_res, txi, samples_grp1, samples_grp2, digits = 4)
}
\arguments{
\item{de_res}{The results from DESeq2::results converted to data.frame.}

\item{txi}{The txi object returned by the import_kallisto function.}

\item{digits}{Integer indicating the number of decimal places}

\item{contrast}{The contrast for the comparison (see ?DESeq2::results).}

\item{design_group}{The column from the design where the contrasts are
located. Default: "group"}

\item{ignoreTxVersion}{Should the transcript version be ignored for anno
mapping. Default: \code{FALSE}.}
}
\value{
A data.frame with the anno and the merged counts values.
}
\description{
The table contains annotation and the DE results. The \code{de_res} input
object must be annotated and correctly formatted (i.e.: from
\code{produce_deliverables}
}
\examples{
txi <- get_demo_txi()
design <- get_demo_design()
dds <- deseq2_analysis(txi, design, ~ group)
de_res <- format_de_results(dds, txi, c("group", "A", "B"))
samples_grp1 <- design[design$group == "A",]$sample
samples_grp2 <- design[design$group == "B",]$sample
de <- format_de_odysse(de_res, txi, samples_grp1, samples_grp2)

}
