% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{format_de_results}
\alias{format_de_results}
\title{Minimal formatting of de results}
\usage{
format_de_results(dds, txi, contrast, keep_stats = TRUE, add_mean_dds = FALSE)
}
\arguments{
\item{dds}{The DESeqDataSet object returned by deseq2_analysis.}

\item{txi}{The txi object returned by the import_kallisto function.}

\item{contrast}{A vector describing the contrasts in the c("<comp_group>",
"<comp1>", "<comp2>") format.}

\item{keep_stats}{Keep baseMean, lfcSE and stat values in the results?
Default: \code{TRUE}.}

\item{add_mean_dds}{Add the mean DESeq normalization value for each group of
the comparison. Default: \code{FALSE}}
}
\value{
A data.frame with the id, ensembl_gene, symbol, entrez_id,
transcript_type, log2FoldChange, pvalue, padj columns.
}
\description{
This function will call DESeq2::results on the dds object and add the
informations from txi$anno using a full join. This means that the id that
were removed in the dds production step by filtering the rows with a small
number of counts will be re-included in the results. In this case, the
log2FoldChange, the pvalue and the padj columns will all be NA.
}
\examples{
txi <- get_demo_txi()
design <- get_demo_design()
dds <- deseq2_analysis(txi, design, ~ group)
de_res <- format_de_results(dds, txi,  c("group", "A", "B"))

}
