% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_de_infos_file}
\alias{get_demo_de_infos_file}
\title{Get demo de_infos file}
\usage{
get_demo_de_infos_file()
}
\value{
The path to the DE infos file
}
\description{
Get demo de_infos file
}
\examples{
de_infos <- get_demo_de_infos_file()

}
