% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_metadata_file}
\alias{get_demo_metadata_file}
\title{Get demo metadata file}
\usage{
get_demo_metadata_file()
}
\value{
The path to the metadata file
}
\description{
Get demo metadata file
}
\examples{
metadata <- get_demo_metadata_file()

}
