% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_pca_infos_file}
\alias{get_demo_pca_infos_file}
\title{Get demo pca infos file}
\usage{
get_demo_pca_infos_file()
}
\value{
The path to the pca infos file
}
\description{
Get demo pca infos file
}
\examples{
pca_infos <- get_demo_pca_infos_file()

}
