% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_volcano_infos_file}
\alias{get_demo_volcano_infos_file}
\title{Get demo volcano_infos file}
\usage{
get_demo_volcano_infos_file()
}
\value{
The path to the DE infos file
}
\description{
Get demo volcano_infos file
}
\examples{
volcano_infos <- get_demo_volcano_infos_file()

}
