% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler2_analysis.R
\name{gprofiler2_analysis}
\alias{gprofiler2_analysis}
\title{Produces a gene profile from DESeq2 results, converts the result for clusterProfiler visualization}
\usage{
gprofiler2_analysis(
  de_res,
  p_threshold = 0.05,
  fc_threshold = 1.5,
  organism = "hsapiens"
)
}
\arguments{
\item{de_res}{either a \code{data.frame} or a list of \code{data.frame} returned by the
\code{DESeq2::results} function.}

\item{p_threshold}{the numeric threshold of the adjusted p-value to be considered as significant.
the value has to be between 0 and 1.
Default: 0.05}

\item{fc_threshold}{the numeric threshold of the fold change to be considered as significant.
the value has to be positive.
Default: 1.5}

\item{organism}{the string that defines corresponding source organism for the gene list.
The organism names are usually constructed by concatenating the first letter of the name
and the family name, e.g human - hsapiens.
Default : "hsapiens"}
}
\value{
a list of two lists. list$gp is the \code{list} returned by \code{gprofiler::gost}.
list$cp is the \code{list} returned by the \code{gprofiler2_convert} function.
}
\description{
Produces a gene profile from DESeq2 results, converts the result for clusterProfiler visualization
}
\examples{
de_res <- get_demo_de_res()
enr <- gprofiler2_analysis(de_res)

}
