% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler2_analysis.R
\name{gprofiler2_convert}
\alias{gprofiler2_convert}
\title{Converts the result of gprofiler2::gost in a compatible object for clusterProfiler}
\usage{
gprofiler2_convert(gp)
}
\arguments{
\item{gp}{the \code{list} returned by the \code{gprofiler::gost}
function wich is a list of two lists.
result contains a data.frame with the results from the hypergeometric test.
meta contains lists of the inputs and the queries for reproductibility.}
}
\value{
cp is a converted \code{list} from the \code{list} returned by the \code{gprofiler::gost}
function and is compatible with the vizualisation function from the code{clusterProfiler} package.
}
\description{
Converts the result of gprofiler2::gost in a compatible object for clusterProfiler
}
