% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_kallisto}
\alias{import_kallisto}
\title{Import quantifications from Kallisto}
\usage{
import_kallisto(filenames, anno, txOut = FALSE, ignoreTxVersion = FALSE)
}
\arguments{
\item{filenames}{Paths to the abundance files.}

\item{anno}{The filename for the annotation in csv (see anno package)
It must have the following columns: id, ensembl_gene, symbol, entrez_id and
transcript_type. Values can be NA except for id and ensembl_gene.}

\item{txOut}{Return counts and abundance at the transcript level. Default:
FALSE}

\item{ignoreTxVersion}{Ignore version of tx. Default = FALSE}
}
\value{
A txi object.
}
\description{
Import quantifications from Kallisto
}
\examples{
abundances <- get_demo_abundance_files()
file_anno <- get_demo_anno_file()
txi <- import_kallisto(abundances, anno = file_anno)

}
