% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plot_pca}
\alias{plot_pca}
\title{Produce a PCA plot from produce_pca_df results}
\usage{
plot_pca(
  res_pca,
  size = 3,
  color = NULL,
  shape = NULL,
  show_names = TRUE,
  title = NULL,
  graph = TRUE,
  legend.position = "right",
  legend.box = "vertical"
)
}
\arguments{
\item{res_pca}{The \code{list} returned from the \code{produce_pca} function}

\item{size}{The size of the points. Default: 3.}

\item{color}{The name of the column in \code{res_pca$coord} to use to color the
points in the PCA. If \code{NULL}, won't add color. Default: \code{NULL}.}

\item{shape}{The name of the column in \code{res_pca$coord} to use to define
the shape of the points in the PCA. If \code{NULL}, won't add change the shape.
Default: \code{NULL}.}

\item{show_names}{Try to add the labels of each point? Will only add the
labels if the points are not too closely clustered together in the PCA.
Default: \code{TRUE}.}

\item{title}{Add a title to the graph? If \code{NULL}, no title is added to
the graph. Otherwise this param must be a \code{character}. Default:
\code{NULL}.}

\item{graph}{Produce the graph. \code{TRUE} or \code{FALSE}. Default:
\code{TRUE}.}

\item{legend.position}{Value for the \code{legend.position} param from
ggplot2. Default: "right".}

\item{legend.box}{Value for the \code{legend.box} param from ggplot2.
Default: "vertical".}
}
\value{
Returns the \code{ggplot} object
}
\description{
Produce a PCA plot from produce_pca_df results
}
\examples{
txi <- get_demo_txi()
res_pca <- produce_pca_df(txi)
p <- plot_pca(res_pca, graph = FALSE)

}
