% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliverables.R
\name{produce_deliverables}
\alias{produce_deliverables}
\title{Produce deliverables}
\usage{
produce_deliverables(
  dir_kallisto,
  anno,
  infos_tables,
  analysis_level = "both",
  file_type = "h5",
  digits = 4,
  ignoreTxVersion = TRUE,
  counts = TRUE,
  add_volcano_labels = NULL,
  report_filename = NULL,
  outdir = "deliverables",
  r_objects = "r_objects",
  force = FALSE,
  ncores = 1,
  verbose = FALSE
)
}
\arguments{
\item{dir_kallisto}{Directory with Kallisto quantifications}

\item{anno}{Path to the anno file matching the reference used for the
Kallisto analysis}

\item{infos_tables}{A \code{list} of infos tables, or path the csv files
that contains these infos. The list may contains the following tables:
pca_infos, de_infos, design_infos, volcano_infos. None of the tables are
mandatory, but only the analysis corresponding to the available tables will
be performed. See the \code{batch_pca}, \code{batch_de},
\code{batch_volcano} and the \code{produce_report} documentation to see the
description of each tables.}

\item{analysis_level}{Should the analysis be done at the "gene", "tx" or
"both" level. Default: "both".}

\item{file_type}{Abundance file format to use (h5 or tsv).}

\item{digits}{Integer indicating the number of decimal places}

\item{ignoreTxVersion}{Should the version number in the IDs be ignored
during import. Default: \code{TRUE}.}

\item{counts}{Save count tables in the <outdir>/counts folder? Default:
\code{TRUE}.}

\item{add_volcano_labels}{Add labels to volcano plots. Must be either
\code{NULL} or a vector of gene symbols. Default: \code{NULL}.}

\item{report_filename}{The name of the report that will be saved in the
\code{outdir/reports/<report_filename>.Rmd}. If \code{NULL}, the report
won't be saved, but the lines will be available in the returned list in the
\code{report} slot. Default: \code{NULL}}

\item{outdir}{A \code{character} string corresponding to the directory to
store the output files (pdf, csv, etc.). The outputs will be in
<outdir>/counts, <outdir>/PCA, <outdir>/DE, <outdir/Volcanos,
<outdir>/report, if their respective tables are present in the
\code{infos_tables}. If \code{NULL}, the results won't be saved on disk.
Default: \code{NULL}.}

\item{r_objects}{A \code{character} string corresponding to the directory
where the rds files will be saved. A standardized data structure will be
used. The expected directories are <r_objects>/txi <r_objects>/counts
<r_objects>/de and <r_objects>/volcanos, each with a "gene/" and "tx/"
folder, except for the txi folder. The folders will only be present if the
respective table is present in the \code{infos_tables}. If \code{NULL}, the
rds files won't be saved on disk. It is recommended to use the
\code{r_objects} param to save files as every steps are completed in case a
problem cause the function to stop. That way, you won't have to reprocess
the steps that completed correctly.  Default: "r_objects".}

\item{force}{Should the files be re-created if they already exists? Default:
\code{FALSE}.}

\item{ncores}{Number of cores to use for de analysis. Default \code{1}.}

\item{verbose}{Print progression of the analysis? Default: \code{FALSE}.}

\item{add_volcanos_labels}{A vector of the symbols to show on the volcano
plots. If \code{NULL}, no symbols will be shown. Default: \code{NULL}.}
}
\value{
Invisibly returns a \code{list} that contains all the results.
}
\description{
The goal of this function is to parse the variaus infos tables and to
produce and save the results in a standardized object and directories
format.
}
\details{
The results will be returned as a list with the following slots:

* txi$gene: The txi object at the gene level.
* txi$tx: The txi object at the transcript level.
* counts$gene: A list of count matrices at the gene level matching the count
               matrices found in the txi object.
* counts$tx: A list of count matrices at the transcript level matching the
             count matrices found in the txi object.
* pca$gene: The list of PCA ggplot objects at the gene level.
* pca$tx: The list of PCA ggplot objects at the transcript level.
* de$gene: The list of DE results in the gene format.
* de$tx: The list of DE results in the transcript format.
* de$volcano$gene: The volcano plots produced at the gene level.
* de$volcano$tx: The volcano plots produced at the transcript level.
* de$report: The lines of the report.
* de$infos_tables: The infos tables used to produce the results.

It's the content of the infos_tables that will determine which deliverables
will be produced. Some tables are mandatory for other steps to be performed.
DE analysis requires de_infos and design_infos tables. Volcano analysis
requires the DE analysis to also be performed beforehand.
}
