% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odysse.R
\name{produce_odysse_format}
\alias{produce_odysse_format}
\title{Save the analysis results in odysse format}
\usage{
produce_odysse_format(
  res,
  outdir,
  prefix,
  use_ruv = FALSE,
  digits = 4,
  force = FALSE,
  ...,
  ncores = 1
)
}
\arguments{
\item{res}{The results obtained from the \code{produce_deliverables}
function or in a similar format.}

\item{outdir}{The directory where the tables will be saved}

\item{prefix}{The value to add at the beginning of each lines}

\item{use_ruv}{Use RUVg normalization? Needs to be pre-computed using the
\code{ruvg_normalization} function. Default: \code{FALSE}.}

\item{digits}{Integer indicating the number of decimal places}

\item{force}{Should the files be re-created if they already exists? Default:
\code{FALSE}.}

\item{...}{Extra params for the format_de_results function.}

\item{ncores}{Number of cores to use for de analysis. Default \code{1}.}
}
\value{
A vector of kallisto abundance filenames
}
\description{
The goal of this function is to produce de deliverables in the Odysse
format. The characteristics of this format are:
    
* counts: A single count file for raw counts, TPM and FPKM. The anno is
included in the file and each count cell contains the 3 types of counts in
the <raw_counts>/<tpm>/<fpkm> format.
* de: There is a file produced for each comparison. The results contains the
annotations and the extra analysis (mean TPM counts and splicing analysis).
There is a distinction between the fold change and the ratio where the ratio
is 2^log2FoldChange and the fold change is symetrical (i.e.: ratios smaller
than 1 are converted to their negative value).
}
\details{
In both cases, the results at the transcripts levels are found in the first
part of the file and the results at the gene levels are found in the second
part of the file.
}
\examples{
abundances <- get_demo_abundance_files()

}
