% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{produce_pca}
\alias{produce_pca}
\title{Deprecated}
\usage{
produce_pca(txi, graph = TRUE, use_ruvg = FALSE)
}
\arguments{
\item{txi}{The \code{txi} object returned by the \code{import_kallisto}
function.}

\item{graph}{Produce the graph. \code{TRUE} or \code{FALSE}. Default:
\code{TRUE}.}

\item{use_ruvg}{Use RUVg normalization? Needs to be pre-computed using the
\code{ruvg_normalization} function. Default: \code{FALSE}.}
}
\description{
Deprecated
}
