% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{produce_pca_df}
\alias{produce_pca_df}
\title{Produce a PCA plot from txi}
\usage{
produce_pca_df(
  txi,
  use_normalisation = "none",
  min_counts = 5,
  metadata = NULL,
  id_metadata = NULL,
  ncp = 2
)
}
\arguments{
\item{txi}{The \code{txi} object returned by the \code{import_kallisto}
function.}

\item{use_normalisation}{What kind of normalisation should be used instead
of TPM? Can be either "none", "ruvg" or "combat". Default: "none"}

\item{min_counts}{The minimal mean number of count (TPM by default; see also
the \code{ruvg_normalization} param) required to keep the feature (gene or
transcript) for the PCA.}

\item{id_metadata}{The colname to use to join the metadata with the
coordinate table. Must contains all the sample names found in the txi. If
\code{NULL} and a metadata table is provided, the first column will be used
to join the tables. Default: \code{NULL}.}

\item{ncp}{Number of component to include in the graph. Default: 2.}

\item{graph}{Produce the graph. \code{TRUE} or \code{FALSE}. Default:
\code{TRUE}.}

\item{metatada}{The metadata to add to the coordinate table to add color or
shape with the results and the \code{plot_pca} function. If there is no
metadata available, keep the default value of \code{NULL}. Value can either
be a csv file or a \code{data.frame}. Default: \code{NULL}.}
}
\value{
Produce the PCA and silently returns a \code{list} with the
coordinates and the x and y labels.
}
\description{
The PCA is produced using the TPM by default (can also be the ruvg counts;
see the \code{ruvg_normalization} param)
}
\examples{
txi <- get_demo_txi()
res_pca <- produce_pca_df(txi)

}
