% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{produce_report}
\alias{produce_report}
\title{Produce report in Rmd format}
\usage{
produce_report(report_infos, report_filename = "report.Rmd", verbose = FALSE)
}
\arguments{
\item{report_infos}{A \code{data.frame} or the path to a csv file that
describes the report to produce.}

\item{report_filename}{The name of the rmarkdown file to create. If
\code{NULL}, the report won't be saved to file. Default: report.Rmd}

\item{verbose}{Print progression? Default: \code{FALSE}.}
}
\value{
Invisibly returns the lines saved to the Rmd file.
}
\description{
The function will build a rmarkdown file based on the the
\code{report_infos} table. The table must contain at least the following
columns: \code{add} and \code{value}. The \code{add} must be one of the
following: text, file or plot. The value will vary depending on the type of
add.
}
\details{
You can also add the \code{id_report} column that will need to contains a
unique identifiers for each line of the report_infos table.

For text, it must be a character string corresponding to the text that
will be directly added to the file specified with \code{report_filename}
file.

For file, it must be a valid path to a filename

For plot, it must be a valid path to a \code{rds} file that contains the
plot to show. The plot must be into a format that can be shown using the
\code{print} function

The \code{extra} column will be used only when the \code{add} value is
"plot" and will be added as is to the code chunk. Be aware that not validity
checks are performed. Make sure you use valid chunk option syntax.

The content of the rmarkdown will be created by parsing each line of the
\code{report_infos} table and adding the requested elements in the same
order as they are found in the table.

There will be no validation that the created rmarkdown file is in the
correct format!
}
