% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{split_de_results}
\alias{split_de_results}
\title{Split differential expression results}
\usage{
split_de_results(
  de_res,
  p_threshold = 0.05,
  fc_threshold = 1.5,
  tpm_threshold = NULL
)
}
\arguments{
\item{de_res}{the \code{data.frame} object returned by the
\code{format_de_results} function.}

\item{p_threshold}{The threshold of p stat to be considered as significant.
Default: 1.5}

\item{fc_threshold}{The threshold of FC to be considered as significant.
Default: 0.05}

\item{tpm_threshold}{The threshold of the mean TPM of the current samples to
be considered as significant}
}
\value{
A list of DE tables.
}
\description{
This function will split the DE results table into 4 elements:
    1) de: The original DE table
    2) signif: All the genes considered statistically differentially
               expressed
    3) up: The significant genes that are up-regulated
    4) down: The significant genes that are down-regulated
}
\details{
To be considered significant, a gene must have a padj (qV) value lower or
equal to the \code{p_threshold} param, an absolute fold change greater or
equal to the \code{fc_threshold} param. Also, if \code{tpm_threshold} is not
\code{NULL}, the average TPM across all samples in the comparison must be
greater or equal to the \code{tpm_threshold} value.

It is important to use the results produced with the
\code{format_de_results} function.
}
\examples{
\dontrun{
txi <- get_demo_txi()
design <- get_demo_design()
dds <- deseq2_analysis(txi, design, ~ group)
de <- format_de_results(dds, txi, c("group", "A", "B"))
split_de <- split_de_results(de)
}

}
