% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txi.R
\name{validate_txi}
\alias{validate_txi}
\title{Validate the txi object content}
\usage{
validate_txi(txi)
}
\arguments{
\item{txi}{a \code{txi} object produced by \code{import_kallisto}.}
}
\value{
Invisibly returns \code{TRUE} if all checks pass.
}
\description{
Validate the txi object content
}
\examples{
txi <- get_demo_txi()
validate_txi(txi) 

}
