% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_metadata.R
\name{wrapper_report_LO}
\alias{wrapper_report_LO}
\title{wrapper report LO}
\usage{
wrapper_report_LO(
  metadata,
  txi,
  outdir,
  pca_subset,
  pca_batch_metadata,
  do_pca = TRUE,
  do_DE = TRUE,
  produce_rmd_report = TRUE,
  render_repport = TRUE,
  extra_count_matrix = NULL,
  volcano_split = NULL,
  volcano_facet_x = NULL,
  volcano_facet_y = NULL,
  custom_parsed_metadata = NULL
)
}
\arguments{
\item{metadata}{(data.frame) sample sheet with at least the columns: ID, Compound, Time, Vehicule, Dose, Cell}

\item{txi}{(list) txi object}

\item{pca_subset}{(character) column of metadata, filter for automated PCA (ex.: Cell type)}

\item{pca_batch_metadata}{(character) extra columns for pca coloring (biological or batch effects)}

\item{do_pca}{(logical) do PCA and included the results in report (default = TRUE)}

\item{do_DE}{(logical) do DE and included the results in report (default = TRUE)}

\item{render_repport}{(logical) render Rmd report}

\item{custom_parsed_metadata}{(list of data.frame) when automatic parsing can be problematic}

\item{report_title}{(character), title in the report}
}
\value{
a list containing pca, volcano and report
}
\description{
From metadata and txi, this wrapper parse the metadata and perform all the PCA and volcano (DE).
It produces a Rmd file which can be rendered.
}
