% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler2_analysis.R
\name{get_demo_de_res}
\alias{get_demo_de_res}
\title{Imports a standard resulting data.frame from DESeq2}
\usage{
get_demo_de_res()
}
\value{
the object returned is a .csv file that countains a data.frame with :
25 rows with a padj < 0.05 (significant) & a abs(log2FoldChange) > log2(1.5)
25 rows with a padj < 0.05 (significant) & a abs(log2FoldChange) < log2(1.5)
25 rows with a padj > 0.05 (not significant) & a abs(log2FolChcange) > log2(1.5)
25 rows with a padj > 0.05 (not significant) & a abs(log2FoldChange) < log2(1.5)
}
\description{
Imports a standard resulting data.frame from DESeq2
}
